move m3u8 urlencoding into its own method with tests and comments
