+++ common/src/test/scala/com/gu/media/util/MediaAtomHelpersTest.scala
@@ -55,6 +55,20 @@ class MediaAtomHelpersTest extends FunSuite with MustMatchers {
     getUser("nodots@guardian.co.uk") must be (User("nodots@guardian.co.uk", None, None))
   }
 
+  test("url-encode self-hosted asset keys to urls") {
+    val asset = SelfHostedAsset(List(
+      VideoSource("url encode me.mp4", "video/mp4"),
+      VideoSource("url encode me.m3u8", "video/m3u8"),
+      VideoSource("2025/08/18/My Title--0653ffba-35f4-4883-b961-3139cdaf6c8b-1.0.m3u8", "video/m3u8"))
+    )
+
+    urlEncodeSources(asset, "https://gu.com/videos") mustBe SelfHostedAsset(List(
+      VideoSource("https://gu.com/videos/url+encode+me.mp4", "video/mp4"),
+      VideoSource("https://gu.com/videos/url+encode+me.m3u8", "video/m3u8"),
+      VideoSource("https://gu.com/videos/2025/08/18/My+Title--0653ffba-35f4-4883-b961-3139cdaf6c8b-1.0.m3u8", "video/m3u8"))
+    )
+  }
+
   private def assets(atom: Atom): Seq[Asset] = {
     atom.data.asInstanceOf[AtomData.Media].media.assets.toSeq
   }
