+++ uploader/src/main/scala/com/gu/media/upload/AddAssetToAtom.scala
@@ -1,15 +1,13 @@
 package com.gu.media.upload
 
-import java.net.URLEncoder
 import java.util.Date
 import com.gu.atom.data.PreviewDynamoDataStore
 import com.gu.atom.publish.PreviewKinesisAtomPublisher
-import com.gu.contentatom.thrift.atom.media.MediaAtom
-import com.gu.contentatom.thrift.{Atom, ContentAtomEvent, EventType, User}
+import com.gu.contentatom.thrift.{Atom, ContentAtomEvent, EventType}
 import com.gu.media.aws.{DynamoAccess, KinesisAccess, UploadAccess}
 import com.gu.media.lambda.LambdaWithParams
 import com.gu.media.logging.Logging
-import com.gu.media.model.{AuditMessage, SelfHostedAsset, YouTubeAsset}
+import com.gu.media.model.{AuditMessage, SelfHostedAsset, VideoAsset, YouTubeAsset}
 import com.gu.media.upload.model.Upload
 import com.gu.media.util.MediaAtomHelpers
 import com.gu.media.util.MediaAtomHelpers._
@@ -54,23 +52,13 @@ class AddAssetToAtom extends LambdaWithParams[Upload, Upload] with DynamoAccess 
     publisher.publishAtomEvent(event).recover { case NonFatal(e) => throw e }
   }
 
-  private def getAsset(upload: Upload) = {
+  private def getAsset(upload: Upload): VideoAsset = {
     upload.metadata.asset match {
       case Some(asset: YouTubeAsset) =>
         asset
 
-      case Some(SelfHostedAsset(sources)) =>
-        SelfHostedAsset(sources.map { source =>
-          val parts = source.src.split("/")
-          parts.length match {
-            case 1 =>
-              source.copy(src = s"$selfHostedOrigin/${URLEncoder.encode(source.src, "UTF-8")}")
-            case _ =>
-              val filename = parts.last
-              val path = parts.dropRight(1).mkString("/")
-              source.copy(src = s"$selfHostedOrigin/$path/${URLEncoder.encode(filename, "UTF-8")}")
-          }
-        })
+      case Some(asset: SelfHostedAsset) =>
+        MediaAtomHelpers.urlEncodeSources(asset, selfHostedOrigin)
 
       case None =>
         throw new IllegalStateException("Missing asset")
